改成支持动态加载

+++ src/cache.js
@@ -1,19 +1,38 @@
-import localForage from "localforage"
 import * as project_package from '../package.json'
 const DB_NAME = 'his_settlements';
 let stores = {};
 let version = project_package.db_version
-let old_version = localStorage.getItem('cc_db_ver')
+let localForage = {}
 
-if (old_version != version) {
-  // 数据库版本升级，整个数据库重置
-  let DBDeleteRequest = indexedDB.deleteDatabase(DB_NAME);
-  DBDeleteRequest.onerror = function(event) {
-    console.log("Error deleting database.");
-  };
-  DBDeleteRequest.onsuccess = function(event) {
-    localStorage.setItem('cc_db_ver', version)
-  };
+if (location.protocol === 'data:') {
+  let _datas = {}
+  localForage.createInstance = function ({
+                                           name,
+                                           storeName
+                                         } = {}) {
+    _datas[storeName] = _datas[storeName] || {}
+    stores[storeName] = {
+      getItem: (k) => _datas[storeName][k] || '',
+      setItem: (k, v) => {
+        _datas[storeName][k] = v
+      }
+    }
+  }
+} else {
+  import("localforage").then(function(module){
+    localForage = module
+    let old_version = localStorage.getItem('cc_db_ver')
+    if (old_version != version) {
+      // 数据库版本升级，整个数据库重置
+      let DBDeleteRequest = indexedDB.deleteDatabase(DB_NAME);
+      DBDeleteRequest.onerror = function(event) {
+        console.log("Error deleting database.");
+      };
+      DBDeleteRequest.onsuccess = function(event) {
+        localStorage.setItem('cc_db_ver', version)
+      };
+    }
+  })
 }
 
 export default {
